Niveles de Usuario, CRUD de Empleados

+++ app/Http/Controllers/RubrosController.php
@@ -11,47 +11,54 @@ class RubrosController extends Controller
 {
     public function index()
     {
-        $rubros = RubrosModel::orderBy('ID_Rubro', 'asc')->get(); //ordena los rubros de forma ascendente
+        $rubros = RubrosModel::where('ID_Rubro', '!=', '1')->orderBy('ID_Rubro', 'asc')->get(); //ordena los rubros de forma ascendente
         $titulo = 'Gestión de Rubros';
 
         return view('rubros', compact('rubros', 'titulo'));
     }
 
     //Creación
-    public function store(Request $request)
+    public function create(Request $request)
     {
-        $request->validate(RubrosModel::rules());
+        if ($request->isMethod('get')) {
+            return view('rubros_form');
+        } else if ($request->isMethod('post')) {
+            $request->validate(RubrosModel::rules());
 
-        RubrosModel::create($request->all());
+            RubrosModel::create($request->all());
 
-        return redirect()->route('rubros.index')
-            ->with('success', 'Rubro creado exitosamente.');
+            return redirect()->route('rubros.index')
+                ->with('success', 'Rubro creado exitosamente.');
+        }
     }
 
     //Editar
     public function update(Request $request, $id)
     {
-        $rubro = RubrosModel::findOrFail($id);
+        if ($id == 1) {
+            return redirect()->route('rubros.index');
+        }
+        if ($request->isMethod('get')) {
+            $rubro = RubrosModel::findOrFail($id);
+            return view('rubros_form', compact('rubro'));
+        } else if ($request->isMethod('put')) {
+            $rubro = RubrosModel::findOrFail($id);
 
-        $request->validate([
-            'Nombre' => [
-                'required',
-                'string',
-                'max:100',
-                'regex:/^[\pL\s]+$/u', // Solo letras y espacios
-                Rule::unique('rubros', 'Nombre')->ignore($rubro->ID_Rubro, 'ID_Rubro')
-            ]
-        ]);
+            $request->validate(RubrosModel::rules($id));
 
-        $rubro->update($request->all());
+            $rubro->update($request->all());
 
-        return redirect()->route('rubros.index')
-            ->with('success', 'Rubro actualizado exitosamente.');
+            return redirect()->route('rubros.index')
+                ->with('success', 'Rubro actualizado exitosamente.');
+        }
     }
 
     //Eliminar
     public function destroy($id)
     {
+        if ($id == 1) {
+            return redirect()->route('rubros.index');
+        }
         $rubro = RubrosModel::findOrFail($id);
         $rubro->delete();
